fix(upgrade_test): workaround for shim-signed for deb only
we are right now running this command to every
upgrade test, but RHEL like distros are not suffering
from this problem, so adding a check, and running
it for Ubuntu only.
Refs: #6054

(cherry picked from commit 9a2da4b0138e65d48ed29135017c60b5eb7dd7a1)

+++ upgrade_test.py
@@ -160,7 +160,8 @@ class UpgradeTest(FillDatabaseData, loader_utils.LoaderUtilsMixin):
 
         InfoEvent(message='Upgrading a Node').publish()
         # because of scylladb/scylla-enterprise#2818 we are for now adding this workaround
-        node.remoter.sudo("apt-get remove shim-signed -y --allow-remove-essential")
+        if node.distro.is_ubuntu:
+            node.remoter.sudo("apt-get remove shim-signed -y --allow-remove-essential")
         node.upgrade_system()
 
         # We assume that if update_db_packages is not empty we install packages from there.
